requirements.txt
pip==21.3.1
torch==1.10.1
torchvision==0.11.2
Flask==2.0.2
Pillow==8.4.0